Make sure all responses are documented
Before this fix, non-200 responses did not need to be documented in
openapi spec, because when the view would raise (for example)
`HTTPForbidden` the response validation would be skipped.

With this fix, if you did not document 403 as a response in openapi
spec you would instead get `openapi_core.schema.responses.exceptions.InvalidResponse:
Unknown response http status: 403`.