Upgrade to openapi-core 0.13.x and add option to JSONify errors
The `openapi-core` library we build on top of is upgraded to `0.13.2`
which brings a complete rewrite of the validation mechanism that is now
based on `jsonschema` library. This manifests as different validation error messages.

Additionally, I moved `openapi_validation_error` from `examples/todoapp` into the main package so it becomes a first-class citizen and people can use it without copy/pasting. You need to explicitly enable it by running `config.pyramid_openapi3_JSONify_errors()`. It's quite opinionated but
easy to override and provide your own, based on your OpenAPI schema.

Closes https://github.com/Pylons/pyramid_openapi3/pull/59
Closes https://github.com/Pylons/pyramid_openapi3/pull/33
Closes https://github.com/Pylons/pyramid_openapi3/pull/43
Closes https://github.com/Pylons/pyramid_openapi3/issues/40
Closes https://github.com/Pylons/pyramid_openapi3/issues/34
Refs https://github.com/p1c2u/openapi-core/issues/160