Raise a warning when request validation causes a response validation failure
The request validator can raise a 400 error if the incoming request fails
the validation implied by the spec. If the spec does not allow a matching
400 response from this endpoint, that error will be lost and replaced with
a validation error for a bad response code. While we already log the relevant
information, this adds a warning to explicitly alert programmers that they
have erred with their API specification.