Making sure all endpoints define at least the minimum common responses
A common pitfall when using this package is the following: you define
an endpoint that can result in 400 Bad Request, but you forget to list 400
in the `responses` section of your endpoint in openapi.yaml. This package
then instead returns 500 Internal Server error, because it keeps the promise
that only defined responses will be allowed (unless you set
`enable_request_validation` to `False`, that is).

With this commit, all endpoints, by default need to have 200, 400 and 500
on the list of `responses` in openapi.yaml, otherwise the app won't start. Additionally, all
endpoints that accept a parameter, also need to have 404 on the list of
`responses`.

You can skip this check by setting `enable_endpoint_validation` to `False`.

Refs https://github.com/Pylons/pyramid_openapi3/issues/22
Refs https://github.com/Pylons/pyramid_openapi3/issues/49#issuecomment-628699131
Refs https://github.com/Pylons/pyramid_openapi3/pull/36